💡 Added comments to all commands
