+++ src/commands/admin/credits/index.js
@@ -6,12 +6,11 @@ const { give, take, set, transfer } = require('./addons');
 
 module.exports = async (interaction) => {
   // Destructure member
-
   const { member } = interaction;
 
   // Check permission
-
   if (!member.permissions.has(Permissions.FLAGS.MANAGE_GUILD)) {
+    // Create embed object
     const embed = {
       title: 'Admin',
       color: config.colors.error,
@@ -19,31 +18,36 @@ module.exports = async (interaction) => {
       timestamp: new Date(),
       footer: { iconURL: config.footer.icon, text: config.footer.text },
     };
+
+    // Send interaction reply
     await interaction.editReply({ embeds: [embed], ephemeral: true });
   }
 
-  // Command handler
-
+  // If subcommand is give
   if (interaction.options.getSubcommand() === 'give') {
     // Execute give addon
-
     await give(interaction);
-  } else if (interaction.options.getSubcommand() === 'take') {
+  }
+
+  // If subcommand is take
+  else if (interaction.options.getSubcommand() === 'take') {
     // Execute take addon
-
     await take(interaction);
-  } else if (interaction.options.getSubcommand() === 'set') {
+  }
+
+  // If subcommand is set
+  else if (interaction.options.getSubcommand() === 'set') {
     // Execute set addon
-
     await set(interaction);
-  } else if (interaction.options.getSubcommand() === 'transfer') {
-    // Execute transfer addon
+  }
 
+  // If subcommand is transfer
+  else if (interaction.options.getSubcommand() === 'transfer') {
+    // Execute transfer addon
     await transfer(interaction);
   }
 
   // Send debug message
-
   await logger.debug(
     `Guild: ${member.guild.id} User: ${member.id} executed /${
       interaction.commandName
