+++ src/commands/credits/addons/balance.js
@@ -1,4 +1,3 @@
-const i18next = require('i18next');
 const config = require('../../../../config.json');
 const logger = require('../../../handlers/logger');
 
@@ -7,15 +6,21 @@ const creditNoun = require('../../../helpers/creditNoun');
 
 module.exports = async (interaction) => {
   try {
+    // Get options
     const user = await interaction.options.getUser('user');
 
+    // Get credit object
     await credits
       .findOne({
         userId: user ? user.id : interaction.user.id,
         guildId: interaction.member.guild.id,
       })
+
+      // If successful
       .then(async (data) => {
+        // If user has no credits
         if (!data) {
+          // Create embed object
           const embed = {
             title: 'Balance',
             description: `${user} has no credits.`,
@@ -24,10 +29,14 @@ module.exports = async (interaction) => {
             footer: { iconURL: config.footer.icon, text: config.footer.text },
           };
 
+          // Send interaction reply
           return interaction.editReply({ embeds: [embed], ephemeral: true });
         }
+
+        // Destructure balance
         const { balance } = data;
 
+        // Create embed object
         const embed = {
           title: 'Balance',
           description: `${user ? `${user} has` : 'You have'} ${creditNoun(
@@ -37,10 +46,16 @@ module.exports = async (interaction) => {
           timestamp: new Date(),
           footer: { iconURL: config.footer.icon, text: config.footer.text },
         };
+
+        // Send interaction reply
         return interaction.editReply({ embeds: [embed], ephemeral: true });
       })
-      .catch(async (err) => logger.error(err));
-  } catch {
-    await logger.error();
+      .catch(async (e) => {
+        // Send debug message
+        await logger.error(e);
+      });
+  } catch (e) {
+    // Send debug message
+    await logger.error(e);
   }
 };
