+++ src/commands/credits/addons/index.js
@@ -1,0 +1,6 @@
+const balance = require('./balance');
+const gift = require('./gift');
+const top = require('./top');
+const work = require('./work');
+
+module.exports = { balance, gift, top, work };
