+++ src/commands/settings/user/index.js
@@ -6,32 +6,30 @@ const { appearance } = require('./addons');
 
 module.exports = async (interaction) => {
   // Destructure member
-
   const { member } = interaction;
 
   // Check permission
-
   if (!member.permissions.has(Permissions.FLAGS.MANAGE_GUILD)) {
+    // Create embed object
     const embed = {
       title: 'Settings',
       color: config.colors.error,
-      description: 'You do not have permission to manage this!',
+      description: `You don't have permission to manage this!`,
       timestamp: new Date(),
       footer: { iconURL: config.footer.icon, text: config.footer.text },
     };
+
+    // Send interaction reply
     await interaction.editReply({ embeds: [embed], ephemeral: true });
   }
 
-  // Command handler
-
+  // If subcommand is appearance
   if (interaction.options.getSubcommand() === 'appearance') {
     // Execute appearance addon
-
     await appearance(interaction);
   }
 
   // Send debug message
-
   await logger.debug(
     `Guild: ${member.guild.id} User: ${member.id} executed /${
       interaction.commandName
