+++ src/events/interactionCreate.js
@@ -6,19 +6,25 @@ const { guilds } = require('../helpers/database/models');
 module.exports = {
   name: 'interactionCreate',
   async execute(interaction) {
+    // Destructure member, client
     const { member, client } = interaction;
 
+    // If interaction is command
     if (interaction.isCommand()) {
+      // Get command from collection
       const command = client.commands.get(interaction.commandName);
 
+      // If command do not exist
       if (!command) return;
 
+      // Create guild if it does not exist already
       await guilds.findOne(
         { guildId: member.guild.id },
         { new: true, upsert: true }
       );
 
       try {
+        // Defer reply
         await interaction.deferReply({
           embeds: [
             {
@@ -35,10 +41,17 @@ module.exports = {
           ],
           ephemeral: true,
         });
+
+        // Execute command
         await command.execute(interaction);
+
+        // Send debug message
         await logger.debug(`Executing command: ${interaction.commandName}`);
       } catch (err) {
+        // Send debug message
         await logger.error(err);
+
+        // Send interaction reply
         await interaction.reply({
           embeds: [
             {
@@ -57,35 +70,5 @@ module.exports = {
         });
       }
     }
-
-    // else if (interaction.isButton()) {
-    //   const button = client.buttons.get(interaction.customId);
-
-    //   try {
-    //     if (!button) {
-    //       await interaction.deferReply();
-    //       await interaction.editReply({ content: `Button not exist: ${interaction.customId}` });
-    //     }
-
-    //     await button.execute(interaction);
-    //     await logger.debug(`Button pressed: ${interaction.customId}`);
-    //   } catch (err) {
-    //     await logger.error(err);
-    //   }
-    // } else if (interaction.isSelectMenu()) {
-    //   const menu = client.menus.get(interaction.customId);
-
-    //   try {
-    //     if (!menu) {
-    //       await interaction.deferReply();
-    //       await interaction.editReply({ content: `Menu not exist: ${interaction.customId}` });
-    //     }
-
-    //     await menu.execute(interaction);
-    //     await logger.debug(`Menu pressed: ${interaction.customId}`);
-    //   } catch (err) {
-    //     await logger.error(err);
-    //   }
-    // }
   },
 };
