--- src/handlers/api.js
@@ -1,8 +1,0 @@
-const axios = require('axios');
-
-const config = require('../../config.json');
-
-module.exports = axios.create({
-  baseURL: config.credits.url,
-  headers: { Authorization: `Bearer ${config.credits.token}` },
-});
